Duke A-Packages: Organize into Packages
Create 2 packages, Task and Storage. Subdivide classes into respective public class file.
