+++ src/main/java/Duke.txt
@@ -1,2 +1,3 @@
-Events t /at 12/12/2019 1800 done
 Deadline s /by 12/12/2019 done
+Deadline t /by 12/12/2019 done
+Events a /at 12/12/2019 Mon
