+++ src/main/java/Task/DeadlineException.java
@@ -1,0 +1,7 @@
+package Task;
+
+public class DeadlineException extends Exception {
+    public DeadlineException(String message) {
+        super(message);
+    }
+}
