+++ src/main/java/Task/EventsException.java
@@ -1,0 +1,7 @@
+package Task;
+
+public class EventsException extends Exception {
+    public EventsException(String message) {
+            super(message);
+    }
+}
