+++ src/main/java/Task/TaskTypes.java
@@ -1,0 +1,7 @@
+package Task;
+
+public enum TaskTypes {
+    Todo,
+    Deadline,
+    Events
+}
