+++ src/main/java/Task/TodoException.java
@@ -1,0 +1,7 @@
+package Task;
+
+public class TodoException extends Exception {
+    public TodoException(String message) {
+        super(message);
+    }
+}
